Scraper cleanup
Enable scraping from files by specifying a ``--fromfiles``
option.  When reading from files, don't stream
the JSON, as the parsing was causing CPU usage to jump to 100%.

Paramaterize the location of cached files and the logfile.

Use ``path.join()`` to build paths instead of string
concatenation.

Remove use of Set from collections, thus eliminating need for
collections as a dependency.

Include data items that are in the state (cdc.txt) dataset, but
not in the GSC scrape, in the final output.