Add dataset for in zone data, clean up merger semantics
Add a new dataset class for the in zone data CSV file and reference
it in the main scraper program.

Update the merger semantics a bit to make it easier to understand
what's going on.  Use 'dataset' instead of 'slug' to represent
the identifier for the datasets that are being merged.  Use
id to instead of key to reference the record ids.

I still need to implement a mechanism for merging data, like the
in zone data, where the rows are sparse.  That is, there might not
be a record for every row in the CDC dataset, but where we can
notify the merger when we've gone through all the rows of data.

Add vows as a development dependency as I want to test the merger
behavior.

Addresses #3685