Add support to merger for optional datasets
The merger was originally designed with the idea that there would be
a corresponding record in each of the merged datasets.  However,
this won't be the case for the "in-zone" dataset.  This commit
adds support for specifying optional datasets for the merger and
then calling a method to tell it when all rows have been processed.

Addresses #3685